code updated

+++ lib/src/features/agent/home/domain/models/agt_profile_model.dart
@@ -59,7 +59,7 @@ class DataData {
   String email;
   bool isBiometricRegistered;
   String walletId;
-  int walletBalance;
+  double walletBalance;
   bool isVerified;
   String phoneNumber;
   String deviceId;
